refactor(congregation): move upcoming events component to main directory
