+++ src/features/congregation/upcoming_events/utils.tsx
